pyproject.toml
[project]
name = "thesunset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "astral>=3.2",
    "boto3>=1.38.42",
    "flask>=3.1.1",
    "ipython>=9.3.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pytz>=2025.2",
]
